cube_sim: _dataLen is a size_t, not a uint16_t
Choosing Spectral Bloom at 32x32 in flat mode segfaulted the simulator.

The shim's allocateData compared and stored the requested length narrowed
to uint16_t:

    if (data && _dataLen >= (uint16_t)len) return true;
    _dataLen = (uint16_t)len;

WLED declares _dataLen `unsigned` and compares `_dataLen >= len` with no
cast, so this was a divergence, not a faithful reproduction - and it
silently wrapped every allocation over 64 KB.

It only bites when a SMALLER allocation precedes a larger one that
truncates below it, which is exactly what toggling flat mode does. Flat
lights all nine face-blocks of the net where the cube lights five, so
Spectral Bloom on a 96x96 net wants 58 KB as a cube and 83 KB flat. 83074
truncates to 17538, which is less than the 58 KB already held, the reuse
test passed, and the effect wrote 24 KB past the end of its buffer.

Allocating the flat size FIRST hides it completely, because calloc gets
the true size_t - which is why the earlier sweep found nothing: it set
flat mode before selecting the effect, and so never produced the shrinking
sequence. The reproduction is select in cube mode, run, then toggle.

Verified: the reproduction segfaults before and passes after; 75 effects
toggled flat both directions at B=16 and B=32 all pass; the two parity
measurements are unchanged by this commit.

+++ cube_sim/shim/wled.h
@@ -306,7 +306,19 @@ class Segment {
 
   uint32_t *pixels = nullptr;          // the frame the renderers read
   uint8_t  *data   = nullptr;          // effect scratch
-  uint16_t  _dataLen = 0;
+  // size_t, and NOT uint16_t. WLED declares this `unsigned` and compares
+  // `_dataLen >= len` with no cast; narrowing it here silently wrapped every
+  // allocation over 64 KB.
+  //
+  // It took a segfault to find, because it only bites when a SMALLER
+  // allocation precedes a larger one that truncates below it. Spectral Bloom on
+  // a 96x96 net wants 58 KB as a cube and 83 KB flat, because flat mode lights
+  // all nine face-blocks where the cube lights five. 83074 truncates to 17538,
+  // which is less than the 58 KB already held, so the reuse test passed and the
+  // effect wrote 24 KB past its buffer. Allocating the flat size FIRST hides it
+  // completely - calloc gets the real size - which is why a fuzz that set flat
+  // mode before selecting the effect found nothing.
+  size_t    _dataLen = 0;
   uint32_t  call = 0, step = 0;
   uint16_t  aux0 = 0, aux1 = 0;
 
@@ -316,11 +328,11 @@ class Segment {
   // and do NOT reset `call` when the requirement shrinks. Effects are written
   // around this, so getting it "cleaner" here would hide real bugs.
   bool allocateData(size_t len) {
-    if (data && _dataLen >= (uint16_t)len) return true;
+    if (data && _dataLen >= len) return true;
     if (data) free(data);
     data = (uint8_t *)calloc(len, 1);
     if (!data) { _dataLen = 0; return false; }
-    _dataLen = (uint16_t)len;
+    _dataLen = len;
     return true;
   }
